requirements.txt
Django
DjangoRestframework
psycopg2-binary
